+++ server/logic.js
@@ -1,5 +1,5 @@
 import bcrypt from 'bcrypt'
-import silo from './util.mjs'
+import silo from './util.js'
 
 const
 {config,util}=silo,
@@ -9,9 +9,9 @@ logic.ext=x=>(x.split('.')||['']).pop().toLowerCase()
 logic.authPwdCheck=async (pwd,hash)=>await bcrypt.compare(pwd,hash)
 logic.authPwdHash=async (pwd,saltRounds=10)=>await bcrypt.hash(pwd,saltRounds)
 
-logic.authLogin=async function(db,username,pwd)
+logic.authLogin=async function(db,user,pwd)
 {
-	const hash=db[username]
+	const hash=db[user]
 	//user does not exist//@todo add random delay to make less obvious
 	return !hash?false:await logic.authPwdCheck(pwd,hash)
 }
